standardise project structure
