+++ BefriendMarlonAndGunther/Config.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.Xna.Framework.Input;
+
+namespace SMAPIMod3
+{
+    class Config
+    {
+        public Keys debugKey { get; set; }
+
+        public Config()
+        {
+            debugKey = Keys.J;
+        }
+    }
+}
