+++ BefriendMarlonAndGunther/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Pathoschild.Stardew.ModBuildConfig" version="2.1.0-beta" targetFramework="net45" />
+</packages>
