--- Class1.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Drawing;
-
-namespace SMAPIMod3
-{
-    class PointCreate : Point
-    {
-    }
-}
